fix concat error when fp16

+++ ppcls/arch/backbone/model_zoo/vision_transformer.py
@@ -302,7 +302,7 @@ class VisionTransformer(nn.Layer):
         # B = x.shape[0]
         B = paddle.shape(x)[0]
         x = self.patch_embed(x)
-        cls_tokens = self.cls_token.expand((B, -1, -1))
+        cls_tokens = self.cls_token.expand((B, -1, -1)).astype(x.dtype)
         x = paddle.concat((cls_tokens, x), axis=1)
         x = x + self.pos_embed
         x = self.pos_drop(x)
